trainer/requirements.txt
pandas>=2.0
scikit-learn>=1.4
psycopg2-binary>=2.9
python-dotenv>=1.0
sqlalchemy>=2.0
numpy>=1.26
